<3tbs, Mummi and bilbo!!
Thomas ONE-MAN absolute <3!!

Mummi late dawn (past 4:10 (AM)) presence contribution!!

bilbo requesting to enter IT room (past 5:10 (AM)BRT) and laying on his tapestry contribution!!

sodapoppin (past 5:10 (AM)BRT)  and friends contribution!!

iAssyrian chatting contribution!!

alfredotigolo, mandrakenk and Okbatgames chatting contribution!!

jitspoe, Endesga and Fearitself streaming contribution!!

tomrandall (past 5:10 (AM)BRT)  streaming Ludum contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Goldbargames streaming contribution!!

Endesga (until 5:10 (AM)BRT)  streaming contribution!!

jitspoe (past 5:10 (AM)BRT) and strager streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

-- hi5 contribution...!!

from matrix (netnode-canada-camilo-account) dedicated server running Windows 10 Pro Pro through OVH.pt!!

+++ draw2d_gdiplus/graphics.cpp
@@ -4840,7 +4840,7 @@ gdi_fallback:
 
       ap(Gdiplus::Point) ppoint(polygon.get_data(), polygon.get_count());
 
-      ppath->AddPolygon(ppoint, polygon.get_count());
+      ppath->AddPolygon(ppoint, (INT) polygon.get_count());
 
       m_pgraphics->SetClip(ppath, CombineModeIntersect);
 
@@ -4856,7 +4856,7 @@ gdi_fallback:
 
       ap(Gdiplus::PointF) ppoint(polygon.get_data(), polygon.get_count());
 
-      ppath->AddPolygon(ppoint, polygon.get_count());
+      ppath->AddPolygon(ppoint, (INT) polygon.get_count());
 
       m_pgraphics->SetClip(ppath, CombineModeIntersect);
